feat: Refactored the code and add logic for cancel and clear all buttons
