--- User Interface/FilterSearchApp.swift
@@ -1,18 +1,0 @@
-//
-//  FilterSearchApp.swift
-//  FilterSearch
-//
-//  Created by Mirabella on 17/07/25.
-//
-
-import SwiftUI
-import SwiftData
-
-//@main
-struct FilterSearchApp: App {
-    var body: some Scene {
-        WindowGroup {
-            MainView()
-        }
-    }
-}
