+++ Venue Viewer/Models/booth_data.json
@@ -6,7 +6,7 @@
       "id": 1,
       "boothName": "Avory",
       "boothNumber": "F01",
-      "Hall": "1",
+      "hall": "1",
       "categories": ["Skincare"],
       "flashSaleSchedule": [],
       "crowdlevel": 0,
@@ -21,7 +21,7 @@
       "id": 2,
       "boothName": "Avoskin",
       "boothNumber": "F02",
-      "Hall": "1",
+      "hall": "1",
       "categories": ["Skincare"],
       "flashSaleSchedule": [],
       "crowdlevel": 0,
@@ -36,7 +36,7 @@
       "id": 3,
       "boothName": "Bali Body",
       "boothNumber": "F03",
-      "Hall": "1",
+      "hall": "1",
       "categories": ["Body Care"],
       "flashSaleSchedule": [],
       "crowdlevel": 0,
@@ -51,7 +51,7 @@
       "id": 4,
       "boothName": "Biore",
       "boothNumber": "F04",
-      "Hall": "1",
+      "hall": "1",
       "categories": ["Cleansing"],
       "flashSaleSchedule": [],
       "crowdlevel": 0,
@@ -66,7 +66,7 @@
       "id": 5,
       "boothName": "Cetaphil",
       "boothNumber": "F05",
-      "Hall": "1",
+      "hall": "1",
       "categories": ["Skincare"],
       "flashSaleSchedule": [],
       "crowdlevel": 0,
@@ -81,7 +81,7 @@
       "id": 6,
       "boothName": "CeraVe",
       "boothNumber": "F06",
-      "Hall": "1",
+      "hall": "1",
       "categories": ["Dermatology"],
       "flashSaleSchedule": [],
       "crowdlevel": 0,
@@ -96,7 +96,7 @@
       "id": 7,
       "boothName": "Drunk Elephant",
       "boothNumber": "F07",
-      "Hall": "1",
+      "hall": "1",
       "categories": ["Luxury"],
       "flashSaleSchedule": [],
       "crowdlevel": 0,
@@ -111,7 +111,7 @@
       "id": 8,
       "boothName": "Emina",
       "boothNumber": "F08",
-      "Hall": "1",
+      "hall": "1",
       "categories": ["Local Brand"],
       "flashSaleSchedule": [],
       "crowdlevel": 0,
@@ -126,7 +126,7 @@
       "id": 9,
       "boothName": "Erha",
       "boothNumber": "F09",
-      "Hall": "1",
+      "hall": "1",
       "categories": ["Dermatology"],
       "flashSaleSchedule": [],
       "crowdlevel": 0,
@@ -141,7 +141,7 @@
       "id": 10,
       "boothName": "Freeman",
       "boothNumber": "F10",
-      "Hall": "1",
+      "hall": "1",
       "categories": ["Mask"],
       "flashSaleSchedule": [],
       "crowdlevel": 0,
@@ -156,7 +156,7 @@
       "id": 11,
       "boothName": "Garnier",
       "boothNumber": "F11",
-      "Hall": "1",
+      "hall": "1",
       "categories": ["Drugstore"],
       "flashSaleSchedule": [],
       "crowdlevel": 0,
