+++ Venue Viewer/Views/Booth/Filter/ActionButtons.swift
@@ -8,15 +8,21 @@
 import SwiftUI
 
 struct ActionButtons: View {
-    @Binding var selectedOptions: Set<String>
-    var selectedCategories: Binding<[Category]>? = nil  // optional
+    //    @Binding var selectedOptions: Set<String>
+    var brandOptions: Binding<Set<String>>?
+    var boothTrafficOptions: Binding<Set<String>>?
+    var categoriesOptions: Binding<Set<String>>?
+    //    var selectedCategories: Binding<[Category]>? = nil  // optional
+    @Environment(\.dismiss) private var dismiss
     
-
+    
     var body: some View {
         HStack(spacing: 16) {
             Button(action: {
-                selectedOptions.removeAll()
-                selectedCategories?.wrappedValue.removeAll()  // hanya jika tersedia
+                brandOptions?.wrappedValue.removeAll()
+                boothTrafficOptions?.wrappedValue.removeAll()
+                categoriesOptions?.wrappedValue.removeAll()
+                //                selectedCategories?.wrappedValue.removeAll()  // hanya jika tersedia
             }) {
                 Text("Clear All")
                     .foregroundColor(Color(.primary))
@@ -29,9 +35,10 @@ struct ActionButtons: View {
                         RoundedRectangle(cornerRadius: 7)
                             .stroke(Color.black, lineWidth: 1))
             }
-
+            
             Button(action: {
-                print("Apply with: \(selectedOptions)")
+//                print("Apply with: \(brandOptions), \(boothTrafficOptions), \(categoriesOptions)")
+                dismiss()
             }) {
                 Text("Apply")
                     .font(.subheadline)
@@ -49,13 +56,13 @@ struct ActionButtons: View {
 }
 
 
-struct ActionButtons_Previews: PreviewProvider {
-    static var previews: some View {
-        ActionButtons(
-            selectedOptions: .constant([]),
-            selectedCategories: .constant([])
-        )
-        .padding()
-        .previewLayout(.sizeThatFits)
-    }
-}
+//struct ActionButtons_Previews: PreviewProvider {
+//    static var previews: some View {
+//        ActionButtons(
+//            selectedOptions: .constant([]),
+////            selectedCategories: .constant([])
+//        )
+//        .padding()
+//        .previewLayout(.sizeThatFits)
+//    }
+//}
