+++ Venue Viewer/Views/Booth/Filter/BoothTraffic.swift
@@ -1,45 +1,43 @@
 //
-//  BrandOffers.swift
+//  BoothTraffic.swift
 //  FilterSearch
 //
 //  Created by Mirabella on 21/07/25.
 //
+
 import SwiftUI
 
-struct BrandOffers: View {
+struct BoothTraffic: View {
+    @Binding var selectedOptions: Set<String>
     let isBottomSheetModal: Bool
-    @State private var selectedOptions: Set<String> = []
-    @State private var selectedCategories: Set<String> = []
-
-       
     
     var body: some View {
-        VStack(alignment: isBottomSheetModal ? .center : .leading, spacing: 16){
-            if(isBottomSheetModal) {
-                Text("Brand Offers")
+        VStack(alignment: isBottomSheetModal ? .center : .leading, spacing: 0) {
+            if(isBottomSheetModal){
+                Text("Booth Traffics")
                     .font(.subheadline)
                     .fontWeight(.semibold)
                     .foregroundColor(.black)
                     .frame(maxWidth: .infinity, alignment: .center)
                     .padding(.top, 40)
+                    .padding(.bottom, 20)
                 Divider()
-                    .padding(.top, 10)
-               
+//                    .padding(.top, 10)
             } else {
 //                Divider()
 //                    .padding(.bottom, 10)
-                Text("Brand Offers")
+                Text("Booth Traffics")
                     .font(.subheadline)
                     .fontWeight(.semibold)
                     .padding(.top, 20)
-                    .padding(.bottom, 10)
-                    
 //                    .padding(.horizontal)
+//                    .padding(.leading, 18)
                     .frame(maxWidth: .infinity, alignment: .leading)
+                
             }
             
             HStack(spacing: 10) {
-                ForEach(["Flash Sale", "Freebies", "Hands On Tester"], id: \.self) { option in
+                ForEach(["High Crowd", "Medium Crowd", "Low Crowd"], id: \.self) { option in
                     Button(action: {
                         if selectedOptions.contains(option) {
                             selectedOptions.remove(option)
@@ -58,33 +56,31 @@ struct BrandOffers: View {
                             )
                             .foregroundColor(.black)
                             .cornerRadius(8)
-                            
                     }
                 }
             }
-     
-//            .padding(.top, 20)
-            .padding(.bottom, 10)
+            .padding(.top, 20)
+            .padding(.bottom, 20)
             
             if(isBottomSheetModal) {
                 Divider()
+                    .padding(.vertical, 10)
                 
                 ActionButtons(
-                    selectedOptions: $selectedOptions
+                    brandOptions: nil,
+                    boothTrafficOptions: nil,
+                    categoriesOptions: $selectedOptions
                 )
-                .padding(.horizontal)
-                .padding(.top, 20)
+//                .padding(.horizontal)
+//                .padding(.top, 20)
+                
             }
-            
         }
-//        .background(Color.red)
-//        .cornerRadius(24)
     }
 }
 
-
-struct BrandOffers_Previews: PreviewProvider {
-    static var previews: some View {
-        BrandOffers(isBottomSheetModal:false)
-    }
-}
+//struct BoothTrafficPreviews: PreviewProvider {
+//    static var previews: some View {
+//        BoothTraffic(isBottomSheetModal: false)
+//    }
+//}
