+++ Venue Viewer/Views/Booth/Filter/CategoriesSheetView.swift
@@ -7,17 +7,12 @@
 
 import SwiftUI
 
-struct Category: Identifiable, Equatable, Hashable {
-    let id = UUID()
-    let name: String
-    let imageName: String
-    let selectedImageName: String
-    let isBottomSheetModal: Bool = false
-}
-
 struct CategoriesSheetView: View {
-    let isBottomSheetModal: Bool
+    @State private var selectedCategories: [Category] = []
     
+    @Binding var selectedOptions: Set<String>
+    
+    let isBottomSheetModal: Bool
     let allCategories: [Category] = [
         .init(name: "Skincare", imageName: "skincare", selectedImageName: "skincare-fill"),
         .init(name: "Make Up", imageName: "makeup", selectedImageName: "makeup-fill"),
@@ -31,9 +26,6 @@ struct CategoriesSheetView: View {
         .init(name: "Salon & Clinic", imageName: "salon", selectedImageName: "salon-fill")
     ]
     
-    @State private var selectedCategories: [Category] = []
-    @State private var selectedOptions: Set<String> = []
-    
     let columns = [
         GridItem(.flexible()),
         GridItem(.flexible()),
@@ -43,20 +35,20 @@ struct CategoriesSheetView: View {
     
     var body: some View {
         VStack{
-         VStack(alignment: isBottomSheetModal ? .center : .leading, spacing: 16) {
+            VStack(alignment: isBottomSheetModal ? .center : .leading, spacing: 16) {
                 if isBottomSheetModal {
                     Text("Categories")
                         .font(.subheadline)
                         .fontWeight(.semibold)
                         .foregroundColor(.black)
                         .frame(maxWidth: .infinity, alignment: .center)
-//                        .padding(.top, 40)
+                    //                        .padding(.top, 40)
                     Divider()
                         .padding(.top, 10)
                 } else {
-                        Text("Categories")
-                            .font(.subheadline)
-                            .fontWeight(.semibold)
+                    Text("Categories")
+                        .font(.subheadline)
+                        .fontWeight(.semibold)
                         .padding(.top, 20)
                         .padding(.horizontal)
                         .frame(maxWidth: .infinity, alignment: .leading)
@@ -76,19 +68,26 @@ struct CategoriesSheetView: View {
                     }
                 }
                 .padding(.horizontal)
-              
+                
                 if(isBottomSheetModal) {
                     Divider()
                     ActionButtons(
-                        selectedOptions: $selectedOptions,
-                        selectedCategories: $selectedCategories
+                        brandOptions: nil,
+                        boothTrafficOptions: nil,
+                        categoriesOptions: $selectedOptions
                     )
                 }
             }
-//            .padding(.horizontal)
+            //            .padding(.horizontal)
         }
         .background(Color.white)
         .cornerRadius(24)
+        .onChange(of: selectedOptions) {
+            newValue in
+                selectedCategories.removeAll { category in
+                    !newValue.contains(category.name)
+                }
+        }
     }
     
     private func toggleCategory(_ category: Category) {
@@ -104,44 +103,6 @@ struct CategoriesSheetView: View {
     }
 }
 
-struct CategoryCard: View {
-    let category: Category
-    let isSelected: Bool
-    let onTap: () -> Void
-    
-    var body: some View {
-        Button(action: onTap) {
-            VStack {
-                Image(isSelected ? category.selectedImageName : category.imageName)
-                    .resizable()
-                    .scaledToFit()
-                    .frame(width: 40, height: 40) 
-                Text(category.name)
-                    .font(.caption)
-                    .lineLimit(2)
-                    .multilineTextAlignment(.center)
-                    .fixedSize(horizontal: false, vertical: true)
-                    .minimumScaleFactor(0.7)
-                    .padding(.bottom, 6)
-                    .foregroundColor(.primary)
-            }
-            .frame(width: 72, height: 72)
-            .padding(5)
-            .background(
-                RoundedRectangle(cornerRadius: 10)
-                    .fill(isSelected ? Color.pink.opacity(0.2) : Color.clear)
-            )
-            .overlay(
-                RoundedRectangle(cornerRadius: 10)
-                    .stroke(isSelected ? Color.pink : Color.gray.opacity(0.3), lineWidth: isSelected ? 2 : 1)
-            )
-            .cornerRadius(10)
-        }
-        .buttonStyle(.plain)
-    }
-}
-
-
-#Preview {
-    CategoriesSheetView(isBottomSheetModal: false)
-}
+//#Preview {
+//    CategoriesSheetView(isBottomSheetModal: false)
+//}
